refactored changeWithTimeDialog
put it in a dedicated class
